Import DSM 7.0-41890 kvmx64 linux-4.4.x.txz Base on Linux Kernal v4.4.180
tar url: https://global.synologydownload.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890/kvmx64/linux-4.4.x.txz
tar md5sum: b1261d255766e43820e021b8a02f8c49
from: https://archive.synology.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890

+++ arch/hexagon/kernel/dma.c
@@ -46,6 +46,7 @@ EXPORT_SYMBOL(dma_supported);
 
 static struct gen_pool *coherent_pool;
 
+
 /* Allocates from a pool of uncached memory that was reserved at boot time */
 
 static void *hexagon_dma_alloc_coherent(struct device *dev, size_t size,
@@ -67,7 +68,7 @@ static void *hexagon_dma_alloc_coherent(struct device *dev, size_t size,
 			panic("Can't create %s() memory pool!", __func__);
 		else
 			gen_pool_add(coherent_pool,
-				pfn_to_virt(max_low_pfn),
+				(unsigned long)pfn_to_virt(max_low_pfn),
 				hexagon_coherent_pool_size, -1);
 	}
